requirements.txt
enum34==1.0
python-registry==1.0.4
wsgiref==0.1.2
